feat: boardClubComment 로그인 인증 추가

+++ server/src/main/java/com/splashzone/boardclubcomment/controller/BoardClubCommentController.java
@@ -1,16 +1,21 @@
 package com.splashzone.boardclubcomment.controller;
 
+import com.splashzone.auth.userdetails.MemberDetails;
 import com.splashzone.boardclubcomment.dto.BoardClubCommentDto;
 import com.splashzone.boardclubcomment.entity.BoardClubComment;
 import com.splashzone.boardclubcomment.mapper.BoardClubCommentMapper;
 import com.splashzone.boardclubcomment.service.BoardClubCommentService;
 import com.splashzone.dto.MultiResponseDto;
 import com.splashzone.dto.SingleResponseDto;
+import com.splashzone.member.entity.Member;
+import com.splashzone.member.service.MemberService;
 import com.splashzone.utils.UriCreator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,6 +23,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import java.net.URI;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/clubcomments")
@@ -27,25 +33,40 @@ public class BoardClubCommentController {
     private final static String BOARD_CLUB_COMMENT_DEFAULT_URL = "/clubcomments";
     private final BoardClubCommentService boardClubCommentService;
     private final BoardClubCommentMapper boardClubCommentMapper;
+    private final MemberService memberService;
 
     @PostMapping
-    public ResponseEntity postBoardClubComment(@Valid @RequestBody BoardClubCommentDto.Post requestBody) {
-        BoardClubComment boardClubComment = boardClubCommentService.createBoardClubComment(
-                boardClubCommentMapper.boardClubCommentPostDtotoBoardClubComment(requestBody));
+    public ResponseEntity postBoardClubComment(Authentication authentication,
+                                               @Valid @RequestBody BoardClubCommentDto.Post requestBody) {
+        UserDetails memberDetails = (MemberDetails) authentication.getPrincipal();
 
-        URI location = UriCreator.createUri(BOARD_CLUB_COMMENT_DEFAULT_URL, boardClubComment.getBoardClubCommentId());
+        Member member = memberService.findMemberByUsername(memberDetails.getUsername());
+
+        BoardClubComment boardClubComment = boardClubCommentMapper.boardClubCommentPostDtotoBoardClubComment(requestBody, member);
+        boardClubComment.setMember(member);
+
+        BoardClubComment postBoardClubComment = boardClubCommentService.createBoardClubComment(boardClubComment);
+
+        URI location = UriCreator.createUri(BOARD_CLUB_COMMENT_DEFAULT_URL, postBoardClubComment.getBoardClubCommentId());
 
         return ResponseEntity.created(location).build();
     }
 
     @PatchMapping("/{club-comment-id}")
-    public ResponseEntity patchBoardClubComment(@PathVariable("club-comment-id") @Positive Long boardClubCommentId,
+    public ResponseEntity patchBoardClubComment(Authentication authentication,
+                                                @PathVariable("club-comment-id") @Positive Long boardClubCommentId,
                                                 @Valid @RequestBody BoardClubCommentDto.Patch requestBody) {
-        BoardClubComment boardClubComment = boardClubCommentService.updateBoardClubComment(
-                boardClubCommentMapper.boardClubCommentPatchDtotoBoardClubComment(requestBody, boardClubCommentId));
+        UserDetails memberDetails = (MemberDetails) authentication.getPrincipal();
+
+        Member member = memberService.findMemberByUsername(memberDetails.getUsername());
+
+        BoardClubComment boardClubComment = boardClubCommentMapper.boardClubCommentPatchDtotoBoardClubComment(requestBody, boardClubCommentId, member);
+        boardClubComment.setMember(member);
+
+        BoardClubComment postBoardClubComment = boardClubCommentService.updateBoardClubComment(boardClubComment);
 
         return new ResponseEntity<>(
-                new SingleResponseDto<>(boardClubCommentMapper.boardClubCommentToBoardClubCommentResponseDto(boardClubComment)),
+                new SingleResponseDto<>(boardClubCommentMapper.boardClubCommentToBoardClubCommentResponseDto(postBoardClubComment)),
                 HttpStatus.OK);
     }
 
@@ -69,8 +90,19 @@ public class BoardClubCommentController {
                 HttpStatus.OK);
     }
 
+
+    //TODO 로그인인증 추가하기!
     @DeleteMapping("/{club-comment-id}")
-    public ResponseEntity deleteBoardClubComment(@PathVariable("club-comment-id") @Positive Long boardClubCommentId) {
+    public ResponseEntity deleteBoardClubComment(Authentication authentication,
+                                                 @PathVariable("club-comment-id") @Positive Long boardClubCommentId) {
+        UserDetails memberDetails = (MemberDetails) authentication.getPrincipal();
+
+        if (!Objects.equals(memberService.findMemberByUsername(memberDetails.getUsername()),
+                boardClubCommentService.findBoardClubComment(boardClubCommentId).getMember())) {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
+
+        //
         boardClubCommentService.deleteBoardClubComment(boardClubCommentId);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
